fix: rename FileCriteriaControls.module.scss to FileCriteriaForm.module.scss
