+++ src/sections/dataset/dataset-files/file-criteria-form/FileCriteriaFilters.tsx
@@ -1,6 +1,6 @@
 import { FileCriteria } from '../../../../files/domain/models/FileCriteria'
 import { FilesCountInfo } from '../../../../files/domain/models/FilesCountInfo'
-import styles from './FileCriteriaControls.module.scss'
+import styles from './FileCriteriaForm.module.scss'
 import { FileCriteriaFilterByType } from './FileCriteriaFilterByType'
 import { FileCriteriaFilterByAccess } from './FileCriteriaFilterByAccess'
 import { FileCriteriaFilterByTag } from './FileCriteriaFilterByTag'
